requirements.txt
python-telegram-bot==20.7
google-generativeai==0.7.2
PyMuPDF==1.24.0
python-docx==1.1.0
weasyprint==61.2
python-dotenv==1.0.1